Extract repository extensions to separate files
