+++ App/Tasks/ReadTasksInteractor.swift
@@ -15,9 +15,9 @@ class ReadTasksInteractor: RepositoryInteractor {
         self.repository.queryTasks(page, completion: { (tasks, error) in
             completion(tasks)
             
-            remoteRepository?.queryTasks(page, completion: { (tasks, error) in
-                
-            })
+//            remoteRepository?.queryTasks(page, completion: { (tasks, error) in
+//                
+//            })
         })
     }
